+++ po/POTFILES.in
@@ -376,6 +376,8 @@ src/template.cpp
 src/terrain.cpp
 src/text.cpp
 src/texture.cpp
+src/titleui/old.cpp
+src/titleui/titleui.cpp
 src/transporter.cpp
 src/version.cpp
 src/visibility.cpp
