+++ src/Makefile.am
@@ -147,6 +147,7 @@ noinst_HEADERS = \
 	terrain.h \
 	text.h \
 	texture.h \
+	titleui/titleui.h \
 	transporter.h \
 	visibility.h \
 	version.h \
@@ -251,6 +252,8 @@ COMMONSOURCES = \
 	terrain.cpp \
 	text.cpp \
 	texture.cpp \
+	titleui/old.cpp \
+	titleui/titleui.cpp \
 	transporter.cpp \
 	version.cpp \
 	visibility.cpp \
