+++ src/multiint.cpp
@@ -153,9 +153,6 @@ static int					teamChooserUp = -1;
 static int					aiChooserUp = -1;
 static int					difficultyChooserUp = -1;
 static int					positionChooserUp = -1;
-static bool				SettingsUp		= false;
-static UBYTE				InitialProto	= 0;
-static W_SCREEN				*psConScreen;
 static SDWORD				dwSelectedGame	= 0;						//player[] and games[] indexes
 static UDWORD				gameNumber;								// index to games icons
 static bool					safeSearch		= false;				// allow auto game finding.
@@ -177,7 +174,6 @@ static void drawReadyButton(UDWORD player);
 
 // Drawing Functions
 static void displayChatEdit(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
-static void intDisplayFeBox(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
 static void displayRemoteGame(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
 static void displayPlayer(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
 static void displayPosition(WIDGET *psWidget, UDWORD xOffset, UDWORD yOffset);
@@ -851,177 +847,6 @@ static void decideWRF()
 	}
 }
 
-
-// ////////////////////////////////////////////////////////////////////////////
-// Connection Options Screen.
-
-void multiOptionsScreenSizeDidChange(unsigned int oldWidth, unsigned int oldHeight, unsigned int newWidth, unsigned int newHeight)
-{
-	if (psConScreen == nullptr) return;
-	psConScreen->screenSizeDidChange(oldWidth, oldHeight, newWidth, newHeight);
-}
-
-/*!
- * Set the server name
- * \param hostname The hostname or IP address of the server to connect to
- */
-void mpSetServerName(const char *hostname)
-{
-	sstrcpy(serverName, hostname);
-}
-
-/**
- * @return The hostname or IP address of the server we will connect to.
- */
-const char *mpGetServerName()
-{
-	return serverName;
-}
-
-static bool OptionsInet()			//internet options
-{
-	psConScreen = new W_SCREEN;
-
-	W_FORMINIT sFormInit;           //Connection Settings
-	sFormInit.formID = 0;
-	sFormInit.id = CON_SETTINGS;
-	sFormInit.style = WFORM_PLAIN;
-	sFormInit.calcLayout = LAMBDA_CALCLAYOUT_SIMPLE({
-		psWidget->setGeometry(CON_SETTINGSX, CON_SETTINGSY, CON_SETTINGSWIDTH, CON_SETTINGSHEIGHT);
-	});
-	sFormInit.pDisplay = intDisplayFeBox;
-	widgAddForm(psConScreen, &sFormInit);
-
-	addMultiBut(psConScreen, CON_SETTINGS, CON_OK, CON_OKX, CON_OKY, MULTIOP_OKW, MULTIOP_OKH,
-	            _("Accept Settings"), IMAGE_OK, IMAGE_OK, true);
-	addMultiBut(psConScreen, CON_SETTINGS, CON_IP_CANCEL, CON_OKX + MULTIOP_OKW + 10, CON_OKY, MULTIOP_OKW, MULTIOP_OKH,
-	            _("Cancel"), IMAGE_NO, IMAGE_NO, true);
-
-	//label.
-	W_LABINIT sLabInit;
-	sLabInit.formID = CON_SETTINGS;
-	sLabInit.id		= CON_SETTINGS_LABEL;
-	sLabInit.style	= WLAB_ALIGNCENTRE;
-	sLabInit.x		= 0;
-	sLabInit.y		= 10;
-	sLabInit.width	= CON_SETTINGSWIDTH;
-	sLabInit.height = 20;
-	sLabInit.pText	= WzString::fromUtf8(_("IP Address or Machine Name"));
-	widgAddLabel(psConScreen, &sLabInit);
-
-
-	W_EDBINIT sEdInit;             // address
-	sEdInit.formID = CON_SETTINGS;
-	sEdInit.id = CON_IP;
-	sEdInit.x = CON_IPX;
-	sEdInit.y = CON_IPY;
-	sEdInit.width = CON_NAMEBOXWIDTH;
-	sEdInit.height = CON_NAMEBOXHEIGHT;
-	sEdInit.pText = serverName;
-	sEdInit.pBoxDisplay = intDisplayEditBox;
-	if (!widgAddEditBox(psConScreen, &sEdInit))
-	{
-		return false;
-	}
-	// auto click in the text box
-	W_CONTEXT sContext;
-	sContext.xOffset	= 0;
-	sContext.yOffset	= 0;
-	sContext.mx			= CON_NAMEBOXWIDTH;
-	sContext.my			= 0;
-	widgGetFromID(psConScreen, CON_IP)->clicked(&sContext);
-
-	SettingsUp = true;
-	return true;
-}
-
-// ////////////////////////////////////////////////////////////////////////////
-// Draw the connections screen.
-bool startConnectionScreen()
-{
-	addBackdrop();										//background
-	addTopForm();										// logo
-	addBottomForm();
-
-	SettingsUp		= false;
-	InitialProto	= 0;
-	safeSearch		= false;
-
-	// don't pretend we are running a network game. Really do it!
-	NetPlay.bComms = true; // use network = true
-
-	addSideText(FRONTEND_SIDETEXT,  FRONTEND_SIDEX, FRONTEND_SIDEY, _("CONNECTION"));
-
-	addMultiBut(psWScreen, FRONTEND_BOTFORM, CON_CANCEL, 10, 10, MULTIOP_OKW, MULTIOP_OKH,
-	            _("Return To Previous Screen"), IMAGE_RETURN, IMAGE_RETURN_HI, IMAGE_RETURN_HI);	// goback buttpn levels
-
-	addTextButton(CON_TYPESID_START + 0, FRONTEND_POS2X, FRONTEND_POS2Y, _("Lobby"), WBUT_TXTCENTRE);
-	addTextButton(CON_TYPESID_START + 1, FRONTEND_POS3X, FRONTEND_POS3Y, _("IP"), WBUT_TXTCENTRE);
-
-	return true;
-}
-
-void runConnectionScreen()
-{
-	W_SCREEN *curScreen = SettingsUp ? psConScreen : psWScreen;
-	WidgetTriggers const &triggers = widgRunScreen(curScreen);
-	unsigned id = triggers.empty() ? 0 : triggers.front().widget->id; // Just use first click here, since the next click could be on another menu.
-
-	switch (id)
-	{
-	case CON_CANCEL: //cancel
-		changeTitleMode(MULTI);
-		bMultiPlayer = false;
-		bMultiMessages = false;
-		break;
-	case CON_TYPESID_START+0: // Lobby button
-		if (LobbyError != ERROR_INVALID)
-		{
-			setLobbyError(ERROR_NOERROR);
-		}
-		changeTitleMode(GAMEFIND);
-		break;
-	case CON_TYPESID_START+1: // IP button
-		OptionsInet();
-		break;
-	case CON_OK:
-		sstrcpy(serverName, widgGetString(psConScreen, CON_IP));
-		if (serverName[0] == '\0')
-		{
-			sstrcpy(serverName, "127.0.0.1");  // Default to localhost.
-		}
-
-		if (SettingsUp == true)
-		{
-			delete psConScreen;
-			psConScreen = nullptr;
-			SettingsUp = false;
-		}
-
-		joinGame(serverName, 0);
-		break;
-	case CON_IP_CANCEL:
-		if (SettingsUp == true)
-		{
-			delete psConScreen;
-			psConScreen = nullptr;
-			SettingsUp = false;
-		}
-		break;
-	}
-
-	widgDisplayScreen(psWScreen);							// show the widgets currently running
-	if (SettingsUp == true)
-	{
-		widgDisplayScreen(psConScreen);						// show the widgets currently running
-	}
-
-	if (CancelPressed())
-	{
-		changeTitleMode(MULTI);
-	}
-}
-
 // ////////////////////////////////////////////////////////////////////////
 // Lobby error reading
 LOBBY_ERROR_TYPES getLobbyError()
@@ -1100,7 +925,7 @@ bool joinGame(const char *host, uint32_t port)
 	setMultiStats(selectedPlayer, playerStats, false);
 	setMultiStats(selectedPlayer, playerStats, true);
 
-	changeTitleMode(MULTIOPTION);
+	changeTitleUI(std::make_shared<WzMultiOptionTitleUI>());
 
 	if (war_getMPcolour() >= 0)
 	{
@@ -2979,8 +2804,7 @@ static void stopJoining()
 		sendLeavingMsg();								// say goodbye
 		NETclose();										// quit running game.
 		bHosted = false;								// stop host mode.
-		widgDelete(psWScreen, FRONTEND_BACKDROP);		// refresh options screen.
-		startMultiOptions(false);
+		changeTitleUI(wzTitleUICurrent);				// refresh options screen.
 		ingame.localJoiningInProgress = false;
 		return;
 	}
@@ -3307,7 +3131,7 @@ static void processMultiopWidgets(UDWORD id)
 		break;
 
 	case MULTIOP_STRUCTLIMITS:
-		changeTitleMode(MULTILIMIT);
+		changeTitleUI(std::make_shared<WzMultiLimitTitleUI>(std::dynamic_pointer_cast<WzMultiOptionTitleUI>(wzTitleUICurrent)));
 		break;
 
 	case MULTIOP_PNAME:
@@ -3671,7 +3495,7 @@ void frontendMultiMessages()
 			recvOptions(queue);
 			ingame.localOptionsReceived = true;
 
-			if (titleMode == MULTIOPTION)
+			if (std::dynamic_pointer_cast<WzMultiOptionTitleUI>(wzTitleUICurrent))
 			{
 				addGameOptions();
 				disableMultiButs();
@@ -3858,7 +3682,7 @@ void frontendMultiMessages()
 	}
 }
 
-void runMultiOptions()
+TITLECODE WzMultiOptionTitleUI::run()
 {
 	static UDWORD	lastrefresh = 0;
 	char                    oldGameMap[128];
@@ -3994,7 +3818,7 @@ void runMultiOptions()
 			// we abort the 'hidetime' on press of a mouse button.
 			if (gameTime - hideTime < MAP_PREVIEW_DISPLAY_TIME && !mousePressed(MOUSE_LMB) && !mousePressed(MOUSE_RMB))
 			{
-				return;
+				return TITLECODE_CONTINUE;
 			}
 			inputLoseFocus();	// remove the mousepress from the input stream.
 			hideTime = 0;
@@ -4027,6 +3851,8 @@ void runMultiOptions()
 		changeTitleMode(GAMEFIND);
 		screen_RestartBackDrop();
 	}
+
+	return TITLECODE_CONTINUE;
 }
 
 static bool loadSettings(const WzString &filename)
@@ -4046,16 +3872,24 @@ static bool loadSettings(const WzString &filename)
 	return true;
 }
 
-bool startMultiOptions(bool bReenter)
+WzMultiOptionTitleUI::WzMultiOptionTitleUI()
 {
+}
+
+void WzMultiOptionTitleUI::start()
+{
+	bool bReenter = performedFirstStart;
+
 	PLAYERSTATS		nullStats;
 	UBYTE i;
+
+	performedFirstStart = true;
 
 	netPlayersUpdated = true;
 
 	addBackdrop();
 	loadMapPreview(false);
-	addTopForm();
+	addTopForm(true);
 
 	if (getLobbyError() != ERROR_INVALID)
 	{
@@ -4108,7 +3942,7 @@ bool startMultiOptions(bool bReenter)
 		if (!hostCampaign((char *)game.name, (char *)sPlayer))
 		{
 			debug(LOG_ERROR, "Failed to host the challenge.");
-			return false;
+			return;
 		}
 		bHosted = true;
 
@@ -4199,8 +4033,6 @@ bool startMultiOptions(bool bReenter)
 			NETsetPlayerConnectionStatus(CONNECTIONSTATUS_NORMAL, NET_ALL_PLAYERS);
 		}
 	}
-
-	return true;
 }
 
 /////////////////////////////////////////////////////////////////////////////////////////
